==> Data_Analysis/Analisis_de_datos/ch_04/3-agregaciones.ipynb <==
# Agregando datos con pandas y numpy

## Sobre los datos
En este cuaderno trabajaremos con 2 conjuntos de datos:
- La cotización de las acciones de Facebook a lo largo de 2018 (obtenida mediante la [`stock_analysis` package](https://github.com/stefmolin/stock-analysis)).
- Datos meteorológicos diarios de Nueva York [National Centers for Environmental Information (NCEI) API](https://www.ncdc.noaa.gov/cdo-web/webservices/v2).

*Nota: El NCEI forma parte de la Administración Nacional Oceánica y Atmosférica (NOAA) y, como puede ver en la URL de la API, este recurso se creó cuando el NCEI se llamaba NCDC. Si la URL de este recurso cambiara en el futuro, puede buscar "NCEI weather API" para encontrar la actualizada.*

## Antecedentes de los datos meteorológicos

Significado de los datos:
- `AWND`: velocidad media del viento
- `PRCP`: precipitación en milímetros
- `SNOW`: nevadas en milímetros
- SNWD`: profundidad de la nieve en milímetros
- TMAX`: temperatura máxima diaria en grados Celsius
- TMIN`: temperatura mínima diaria en grados Celsius

## Setup

import numpy as np
import pandas as pd

fb = pd.read_csv('data/fb_2018.csv', index_col='date', parse_dates=True).assign(
    trading_volume=lambda x: pd.cut(x.volume, bins=3, labels=['low', 'med', 'high'])
)
fb.head()

weather = pd.read_csv('data/weather_by_station.csv', index_col='date', parse_dates=True)
weather.head()

Antes de sumergirnos en cualquier cálculo, vamos a asegurarnos de que `pandas` no pondrá las cosas en notación científica. Modificaremos el formato de visualización de los flotantes. El formato que aplicaremos es `.2f`, que proporcionará al flotante 2 dígitos después del punto decimal:

pd.set_option('display.float_format', lambda x: '%.2f' % x)

## Resumiendo DataFrames
Aprendimos a utilizar `agg()` en la sección [`2-dataframe_operations.ipynb`](./2-dataframe_operations.ipynb) cuando aprendimos acerca de los cálculos de ventanas; sin embargo, podemos llamar a esto en el marco de datos directamente para agregar su contenido en una sola serie:

fb.agg({
    'open': np.mean, 
    'high': np.max, 
    'low': np.min, 
    'close': np.mean, 
    'volume': np.sum
})

Podemos utilizarlo para hallar el total de nevadas y precipitaciones registradas en Central Park en 2018:

weather.query('station == "GHCND:USW00094728"')\
    .pivot(columns='datatype', values='value')[['SNOW', 'PRCP']]\
    .sum()

Esto equivale a pasar `'sum'` a `agg()`:

weather.query('station == "GHCND:USW00094728"')\
    .pivot(columns='datatype', values='value')[['SNOW', 'PRCP']]\
    .agg('sum')

Tenga en cuenta que no estamos limitados a proporcionar una única agregación por columna. Podemos pasar una lista y obtendremos un marco de datos en lugar de una serie. Los valores nulos se colocan donde no tenemos un resultado de cálculo para mostrar:

fb.agg({
    'open': 'mean',
    'high': ['min', 'max'],
    'low': ['min', 'max'],
    'close': 'mean'
})

## Usando `groupby()`
A menudo no queremos agregar todo el marco de datos, sino grupos dentro de él. Para ello, podemos ejecutar `groupby()` antes de la agregación. Si agrupamos por la columna `trading_volume`, obtendremos una fila por cada uno de los valores que tome:

fb.groupby('trading_volume').mean()

Después de llamar a `groupby()`, todavía podemos seleccionar columnas para la agregación:

fb.groupby('trading_volume')['close'].agg(['min', 'max', 'mean'])

Todavía podemos proporcionar un diccionario especificando las agregaciones a realizar, pero pasando una lista para una columna resultará en un índice jerárquico para las columnas:

fb_agg = fb.groupby('trading_volume').agg({
    'open': 'mean',
    'high': ['min', 'max'],
    'low': ['min', 'max'],
    'close': 'mean'
})
fb_agg

El índice jerárquico en las columnas tiene el siguiente aspecto:

fb_agg.columns

Utilizando una comprensión de lista, podemos unir los niveles (en una tupla) con un `_` en cada iteración:

fb_agg.columns = ['_'.join(col_agg) for col_agg in fb_agg.columns]
fb_agg.head()

Podemos agrupar en valores en el índice si le decimos a `groupby()`, que `nivel` usar:

weather.loc['2018-10'].query('datatype == "PRCP"')\
    .groupby(level=0).mean().head().squeeze()

También podemos crear un objeto `Grouper`, que también puede enrollar las fechas en el índice. Aquí, encontramos la precipitación total trimestral por estación:

weather.query('datatype == "PRCP"').groupby(
    ['station_name', pd.Grouper(freq='Q')]
).sum().unstack().sample(5, random_state=1)

Tenga en cuenta que podemos utilizar `filter()` para excluir algunos grupos de la agregación. En este caso, sólo mantenemos los grupos cuyos nombres terminan en "NY US" en el atributo `name` del grupo, que en este caso es el nombre de la estación:

weather.groupby('station_name').filter( # station names with "NY US" in them
    lambda x: x.name.endswith('NY US')
).query('datatype == "SNOW"').groupby('station_name').sum().squeeze() # agregar y hacer una serie (squeeze)

Veamos cuáles son los meses con más precipitaciones. En primer lugar, tenemos que agrupar por días y calcular la media de las precipitaciones en todas las estaciones. Luego podemos agrupar por meses y sumar la precipitación resultante. Usamos `nlargest()` para obtener los 5 meses con mayor precipitación:

weather.query('datatype == "PRCP"')\
    .groupby(level=0).mean()\
    .groupby(pd.Grouper(freq='M')).sum().value.nlargest()

Quizás el resultado anterior fue sorprendente. El refrán dice que "las lluvias de abril traen las flores de mayo"; sin embargo, abril no estaba entre los 5 primeros (ni tampoco mayo). La nieve contará para las precipitaciones, pero eso no explica por qué los meses de verano están por encima de abril. Busquemos los días que representaron un gran porcentaje de la precipitación en un mes determinado.

Para ello, tenemos que calcular la precipitación media diaria en todas las estaciones y, a continuación, hallar el total por mes. Este será el denominador. Sin embargo, para dividir los valores diarios por el total de su mes, necesitaremos una serie de dimensiones iguales. Esto significa que tendremos que utilizar `transform()`:

weather.query('datatype == "PRCP"')\
    .rename(dict(value='prcp'), axis=1)\
    .groupby(level=0).mean()\
    .groupby(pd.Grouper(freq='M'))\
    .transform(np.sum)['2018-01-28':'2018-02-03']

Observe que tenemos el mismo valor repetido para cada día del mes al que pertenece. Esto nos permitirá calcular el porcentaje de la precipitación mensual que se produjo cada día y, a continuación, extraer los valores más grandes:

weather\
    .query('datatype == "PRCP"')\
    .rename(dict(value='prcp'), axis=1)\
    .groupby(level=0).mean()\
    .assign(
        total_prcp_in_month=lambda x: \
            x.groupby(pd.Grouper(freq='M')).transform(np.sum),
        pct_monthly_prcp=lambda x: \
            x.prcp.div(x.total_prcp_in_month)
    )\
    .nlargest(5, 'pct_monthly_prcp')

`transform()` también se puede utilizar en dataframes. Podemos utilizarlo para normalizar fácilmente los datos:

fb[['open', 'high', 'low', 'close']]\
    .transform(lambda x: (x - x.mean()).div(x.std()))\
    .head()

## Tablas dinámicas y tablas cruzadas
Vimos pivotes en [`ch_03/4-reshaping_data.ipynb`](../ch_03/4-reshaping_data.ipynb); sin embargo, no pudimos proporcionar ninguna agregación. Con `pivot_table()`, obtenemos la media por defecto. En su forma más simple, proporcionamos una columna para colocar a lo largo de las columnas:

fb.pivot_table(columns='trading_volume')

Al colocar el volumen de negociación en el índice, obtenemos la transposición:

fb.pivot_table(index='trading_volume')

Con `pivot()`, tampoco hemos podido manejar índices multinivel o índices con valores repetidos. Por esta razón no hemos podido poner los datos meteorológicos en el formato ancho. El método `pivot_table()` resuelve este problema:

weather.reset_index().pivot_table(
    index=['date', 'station', 'station_name'], 
    columns='datatype', 
    values='value',
    aggfunc='median'
).reset_index().tail()

Podemos utilizar la función `pd.crosstab()` para crear una tabla de frecuencias. Por ejemplo, si queremos ver cuántos días de negociación de bajo, medio y alto volumen han tenido las acciones de Facebook cada mes, podemos utilizar crosstab:

pd.crosstab(
    index=fb.trading_volume,
    columns=fb.index.month,
    colnames=['month'] # name the columns index
)

Podemos normalizar con los totales de fila o columna con el parámetro `normalize`. Esto muestra el porcentaje del total:

pd.crosstab(
    index=fb.trading_volume,
    columns=fb.index.month,
    colnames=['month'],
    normalize='columns'
)

Si queremos realizar un cálculo que no sea contar la frecuencia, podemos pasar la columna sobre la que realizar el cálculo a `values` y la función a utilizar a `aggfunc`:

pd.crosstab(
    index=fb.trading_volume,
    columns=fb.index.month,
    colnames=['month'],
    values=fb.close,
    aggfunc=np.mean
)

También podemos obtener subtotales de filas y columnas con el parámetro `margins`. Contemos el número de veces que cada estación registró nieve por mes e incluyamos los subtotales:

snow_data = weather.query('datatype == "SNOW"')
pd.crosstab(
    index=snow_data.station_name,
    columns=snow_data.index.month,
    colnames=['month'],
    values=snow_data.value,
    aggfunc=lambda x: (x > 0).sum(),
    margins=True, # show row and column subtotals
    margins_name='total observations of snow' # name the subtotals
)

<hr>
<div>
    <a href="./2-operaciones_con_dataframe.ipynb">
        <button>&#8592; Notebook Anterior</button>
    </a>
    <a href="./4-time_series.ipynb">
        <button style="float: right;">Proximo Notebook &#8594;</button>
    </a>
</div>
<hr>
